Add files via upload
In this task, we basically used a new approach to generate an image classification model (DNN). 
DNN stands for Deep Neural Network, A subset of machine learning where a model creates multiple layers of input, processing and output and hence (in this task) gives us the required solution or it classifies based on given training data, We trained this model by giving almost 1000 male images and same goes for female. And trained the model, After training we tested this model to check its accuracy and got a 78% true classification rate.
if we talk about the barriers or difficulties while writing this code or generating this model or solving this problem, the first difficulty was of syntax in code, it was relatively a bit difficult as compared to machine learning algorithms, Secomdly, it took abit longer time than Machine leaning models.
In Deployment i was a bit stuck, nevertheless, with the help of internet it was sorted, overall, No such difficulties of unsolveable hurdles were on the way,
But accuracy was not obtained as expected and we were given instructions not to add multiple layers so the accuracy cant be improved.
[Inxol Technologies task 2.docx](https://github.com/Seemab125/AI-Internship/files/12440606/Inxol.Technologies.task.2.docx)